rename UpdaterFunction.h from UpdaterTemplate.h and add init()
